refactor(partners): move shared helpers + find-or-create into modules/shared
