+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/save-my-partner-links.logic-function.ts
@@ -2,7 +2,8 @@ import { type CoreApiClient } from 'twenty-client-sdk/core';
 import { defineLogicFunction, type RoutePayload } from 'twenty-sdk/define';
 import { z } from 'zod';
 
-import { isHttpUrl } from './http-url';
+import { isHttpUrl } from 'src/modules/shared/utils/http-url.util';
+
 import { buildReconcilePlan } from './reconcile-children';
 import { buildAppClient, errorResponse, failureResponse, resolvePartnerFromRequest } from './resolve-partner-from-request';
