+++ packages/twenty-apps/internal/twenty-partners/src/logic-functions/submit-client-brief.logic-function.ts
@@ -3,11 +3,10 @@ import { defineLogicFunction } from 'twenty-sdk/define';
 import { z } from 'zod';
 
 import { readSecretGuardedEvent } from 'src/modules/shared/http/read-secret-guarded-event';
-
 import {
   findOrCreateCompanyByName,
   findOrCreatePersonByEmail,
-} from './find-or-create-company-and-person';
+} from 'src/modules/shared/services/find-or-create-company-and-person.service';
 
 export const SUBMIT_CLIENT_BRIEF_LOGIC_FUNCTION_ID =
   'a8f3c2e1-9b4d-4a7f-8c6e-1d2f3a4b5c6d';
