+++ packages/twenty-apps/internal/twenty-partners/src/modules/shared/graphql/mutations/create-company.ts
@@ -1,0 +1,7 @@
+import type { CoreApiClient } from 'twenty-client-sdk/core';
+
+export function createCompany(client: CoreApiClient, name: string) {
+  return client.mutation({
+    createCompany: { __args: { data: { name } }, id: true },
+  });
+}
