+++ packages/twenty-apps/internal/twenty-partners/src/modules/shared/services/find-or-create-company-and-person.service.ts
@@ -1,17 +1,17 @@
 import type { CoreApiClient } from 'twenty-client-sdk/core';
 
+import { createCompany } from 'src/modules/shared/graphql/mutations/create-company';
+import { createPerson } from 'src/modules/shared/graphql/mutations/create-person';
+import { findCompanyByName } from 'src/modules/shared/graphql/queries/find-company-by-name';
+import { findPersonByEmail } from 'src/modules/shared/graphql/queries/find-person-by-email';
+
 export async function findCompanyIdByExactName(
   client: CoreApiClient,
   companyName: string,
 ): Promise<string | undefined> {
   const name = companyName.trim();
 
-  const lookup = await client.query({
-    companies: {
-      __args: { filter: { name: { eq: name } }, first: 1 },
-      edges: { node: { id: true } },
-    },
-  });
+  const lookup = await findCompanyByName(client, name);
 
   return lookup.companies?.edges?.[0]?.node?.id;
 }
@@ -24,9 +24,7 @@ export async function findOrCreateCompanyByName(
   if (existing !== undefined) return existing;
 
   const name = companyName.trim();
-  const result = await client.mutation({
-    createCompany: { __args: { data: { name } }, id: true },
-  });
+  const result = await createCompany(client, name);
   const id = result.createCompany?.id;
   if (id === undefined) throw new Error('createCompany did not return an id');
   return id;
@@ -38,12 +36,7 @@ export async function findPersonIdByPrimaryEmail(
 ): Promise<string | undefined> {
   const primaryEmail = email.trim();
 
-  const lookup = await client.query({
-    people: {
-      __args: { filter: { emails: { primaryEmail: { eq: primaryEmail } } }, first: 1 },
-      edges: { node: { id: true } },
-    },
-  });
+  const lookup = await findPersonByEmail(client, primaryEmail);
 
   return lookup.people?.edges?.[0]?.node?.id;
 }
@@ -66,18 +59,7 @@ export async function findOrCreatePersonByEmail(
   const existing = await findPersonIdByPrimaryEmail(client, email);
   if (existing !== undefined) return existing;
 
-  const result = await client.mutation({
-    createPerson: {
-      __args: {
-        data: {
-          name: { firstName, lastName },
-          emails: { primaryEmail: email },
-          companyId: input.companyId,
-        },
-      },
-      id: true,
-    },
-  });
+  const result = await createPerson(client, { email, firstName, lastName, companyId: input.companyId });
   const id = result.createPerson?.id;
   if (id === undefined) throw new Error('createPerson did not return an id');
   return id;
