+++ packages/twenty-apps/internal/twenty-partners/src/modules/shared/utils/http-url.util.ts
