+++ packages/twenty-apps/internal/twenty-partners/src/modules/shared/utils/strip-markdown.util.ts
